refactor(DevToolbox): Extract login functionality to useTestLoginAs hook
